path stuff
Path stuff is always a pain. I've decided we'll want to allow people to create
new "cookbooks" (sqlite databases) and we'll want to populate these new dbs
with the contents of data/create.sql. But of course after installation, what is
that? And then, if you code for the installed case what about the uninstalled
(testing) case? This code deals with that in I think the best way. Apparently
rubygems doesn't use Config::CONFIG['datadir'] so that's another bridge to
cross if/when we gemify it. I'd like to think that most people will be happy to
download, untar, and run, though. For now, run it using "rake run" (or leave
off the run)

darcs-hash:20051014184157-28bda-be17d7ae3d57eb3e724705bbc43a92be5f9c6c68.gz