ingredients table
This is a big one. Lots of GUI progress, not the least of which is that
ingredients table working just exactly as we had always dreamed. It's
beautiful. It might make you cry.

Lots of interesting tricks in this one, including the obligatory
'class << something' trick. Lots of mind-numbing nitty-gritty details that are
likely to hide an infestation of UI bugs, too.

darcs-hash:20051030220251-28bda-6f52c22a11359ada6071629ee637c3ab4ce676e0.gz